Added fading in/out for Tracker when logger messages are displayed
