+++ src/renderer/main/grid-layout/Tracker.svelte
@@ -2,9 +2,11 @@
   import { appSettings } from "../../runtime/app-helper.store";
   import mixpanel from "mixpanel-browser";
   import TooltipSetter from "../user-interface/tooltip/TooltipSetter.svelte";
+  import { fly } from "svelte/transition";
 
   let classProps;
   export { classProps as class };
+  export let visible = true;
 
   const options = [
     {
@@ -30,23 +32,27 @@
   }
 </script>
 
-<div
-  class="{classProps} flex flex-row gap-4 items-center bg-primary py-2 px-3 flex-wrap justify-center rounded-lg"
->
-  <span class="text-white">Interaction Tracking:</span>
-  <div class="flex flex-row gap-2 items-center">
-    {#each options as { value, label, tooltip_key }}
-      <button
-        class:selected={value === $appSettings.changeOnEvent}
-        on:click={() => handleClick(value)}
-        class="w-24 rounded bg-select text-white hover:bg-select-saturate-10 relative py-1"
-      >
-        <span>{label}</span>
-        <TooltipSetter key={tooltip_key} />
-      </button>
-    {/each}
+{#if visible}
+  <div
+    in:fly={{ x: -10, delay: 100 }}
+    out:fly={{ x: 10, delay: 0 }}
+    class="{classProps} flex flex-row gap-4 items-center bg-primary py-2 px-3 flex-wrap justify-center rounded-lg"
+  >
+    <span class="text-white">Interaction Tracking:</span>
+    <div class="flex flex-row gap-2 items-center">
+      {#each options as { value, label, tooltip_key }}
+        <button
+          class:selected={value === $appSettings.changeOnEvent}
+          on:click={() => handleClick(value)}
+          class="w-24 rounded bg-select text-white hover:bg-select-saturate-10 relative py-1"
+        >
+          <span>{label}</span>
+          <TooltipSetter key={tooltip_key} />
+        </button>
+      {/each}
+    </div>
   </div>
-</div>
+{/if}
 
 <style>
   button.selected {
